move unit tests to separate folders
