+++ test/algorithms/classifiers/qsvm/test_qsvm.py
