+++ test/algorithms/classifiers/sklearn_svm/test_sklearn_svm.py
