+++ test/algorithms/classifiers/test_vqc.py
