+++ test/algorithms/distribution_learners/test_qgan.py
