+++ test/datasets/test_digits.py
@@ -24,15 +24,18 @@ class TestDigits(QiskitMachineLearningTestCase):
     def test_digits(self):
         """Digits test."""
 
-        input_file = self.get_resource_path('sample_train.digits')
+        input_file = self.get_resource_path('sample_train.digits',
+                                            'datasets')
         with open(input_file) as file:
             sample_train_ref = json.load(file)
 
-        input_file = self.get_resource_path('training_input.digits')
+        input_file = self.get_resource_path('training_input.digits',
+                                            'datasets')
         with open(input_file) as file:
             training_input_ref = json.load(file)
 
-        input_file = self.get_resource_path('test_input.digits')
+        input_file = self.get_resource_path('test_input.digits',
+                                            'datasets')
         with open(input_file) as file:
             test_input_ref = json.load(file)
